Extract progress token parsing
