+++ src/main/java/com/amannmalik/mcp/server/McpServer.java
@@ -100,7 +100,6 @@ import com.amannmalik.mcp.util.ProgressNotification;
 import com.amannmalik.mcp.util.ProgressToken;
 import com.amannmalik.mcp.util.ProgressTracker;
 import com.amannmalik.mcp.validation.InputSanitizer;
-import com.amannmalik.mcp.validation.MetaValidator;
 import com.amannmalik.mcp.validation.SchemaValidator;
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
@@ -463,16 +462,7 @@ public final class McpServer implements AutoCloseable {
     }
 
     private ProgressToken parseProgressToken(JsonObject params) {
-        if (params == null || !params.containsKey("_meta")) return null;
-        JsonObject meta = params.getJsonObject("_meta");
-        MetaValidator.requireValid(meta);
-        if (!meta.containsKey("progressToken")) return null;
-        var val = meta.get("progressToken");
-        return switch (val.getValueType()) {
-            case STRING -> new ProgressToken.StringToken(InputSanitizer.requireClean(meta.getString("progressToken")));
-            case NUMBER -> new ProgressToken.NumericToken(meta.getJsonNumber("progressToken").longValue());
-            default -> throw new IllegalArgumentException("progressToken must be a string or number");
-        };
+        return ProgressCodec.fromMeta(params);
     }
 
     private boolean allowed(Annotations ann) {
